Fixed bug at adding element

+++ mem.h
@@ -1,9 +1,10 @@
 #ifndef __MEMORIZE
 #define __MEMORIZE
-void *getMem(unsigned);
-void *addMem(void *, unsigned);
+#include <stdlib.h>
+void *getMem(size_t);
+void *addMem(void *, size_t);
 void *putItem(void *, unsigned long long);
-unsigned long long getItem(void *, unsigned);
-unsigned size(void *);
-unsigned filled(void *);
+unsigned long long getItem(void *, size_t);
+size_t size(void *);
+size_t filled(void *);
 #endif
